Project_Weather_Station/web_server/requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset_normalizer==3.4.4
click==8.3.1
colorama==0.4.6
flask==3.1.2
flask_cors==6.0.2
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.3
pip==25.3
pyserial==3.5
python_dotenv==1.2.1
requests==2.32.5
urllib3==2.6.2
werkzeug==3.1.4
python-dotenv
jaydebeapi
JPype1
pandas 
scikit-learn 
numpy